+++ activerecord/test/cases/validations/uniqueness_validation_test.rb
@@ -58,6 +58,14 @@ class UniquenessValidationTest < ActiveRecord::TestCase
     assert t2.save, "Should now save t2 as unique"
   end
 
+  def test_validate_uniqueness_with_alias_attribute
+    Topic.alias_attribute :new_title, :title
+    Topic.validates_uniqueness_of(:new_title)
+
+    topic = Topic.new(new_title: 'abc')
+    assert topic.valid?
+  end
+
   def test_validates_uniqueness_with_nil_value
     Topic.validates_uniqueness_of(:title)